third_party/blink/web_tests/platform/generic/html/grouping_content/figure-element-expected.txt

Various tests for the figure element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<figure> default styling:
PASS getStyleValue("figure0","display") is "block"
PASS getStyleValue("figure0","margin-top") is emSize
PASS getStyleValue("figure0","margin-right") is "40px"
PASS getStyleValue("figure0","margin-bottom") is emSize
PASS getStyleValue("figure0","margin-left") is "40px"
<figure> closes <p>:
PASS figure1.parentNode.nodeName == "p" is false
<p> does not close <figure>:
PASS p1.parentNode.nodeName is "FIGURE"
<figure> can be nested inside <figure> or <footer>:
PASS figure3.parentNode.id is "figure2"
PASS figure5.parentNode.id is "footer1"
Residual style:
PASS getStyleValue("figure4","font-weight") is "700"
PASS getStyleValue("span1","font-weight") is "700"
PASS successfullyParsed is true

TEST COMPLETE
